# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trueblue",
        "description": "Trueblue OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/v1/carts/{id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "operationId": "cart.getCart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "rowId": {
                                                        "type": "string",
                                                        "example": "4c13f596efdd61467998d9100b0a0ca1"
                                                    },
                                                    "id": {
                                                        "type": "string",
                                                        "example": "9ymo1knaAZ"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "iPad Mini 4"
                                                    },
                                                    "qty": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": "100"
                                                    },
                                                    "tax": {
                                                        "type": "number",
                                                        "example": "0"
                                                    },
                                                    "subtotal": {
                                                        "type": "number",
                                                        "example": "100"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "subtotal": {
                                            "type": "number",
                                            "example": "100.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "operationId": "cart.createCart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "operationId": "cart.emptyCart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/carts": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "operationId": "cart.createCart",
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "identifier": {
                                            "type": "string",
                                            "example": "nA3RdgxGRrA4HvEeKcLARNfmunAewf6R"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/carts/{id}/{itemId}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "operationId": "cart.createCart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "operationId": "cart.updateItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/categories/{id}/products": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "category.listProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/me/favourites/products": {
            "get": {
                "tags": [
                    "Me"
                ],
                "operationId": "me.listFavourites",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/me": {
            "get": {
                "tags": [
                    "Me"
                ],
                "operationId": "me.showMe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Me"
                ],
                "operationId": "me.updateMe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Peter Chan"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "me@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "title": {
                        "type": "string",
                        "example": "iPad Mini 4"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "price": {
                        "type": "number",
                        "example": "100"
                    },
                    "rating": {
                        "type": "number",
                        "example": "4"
                    },
                    "stock": {
                        "type": "number",
                        "example": "0"
                    },
                    "hasStock": {
                        "type": "boolean",
                        "example": false
                    },
                    "isVip": {
                        "type": "boolean",
                        "example": false
                    },
                    "isVariant": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User",
                "properties": {
                    "id": {
                        "schema": "User",
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "name": {
                        "schema": "User",
                        "type": "string",
                        "example": "Peter Pan"
                    },
                    "email": {
                        "schema": "User",
                        "type": "string",
                        "example": "me@example.com",
                        "nullable": true
                    },
                    "isVip": {
                        "schema": "User",
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}